Refactor extract TuPrologHelpers from PrologGameKnowledge.
